+++ activerecord/CHANGELOG.md
@@ -1,3 +1,8 @@
+*   `ActiveRecord::Store` works together with PG `hstore` columns.
+    Fixes #12452.
+
+    *Yves Senn*
+
 *   Fix bug where `ActiveRecord::Store` used a global `Hash` to keep track of
     all registered `stored_attributes`. Now every subclass of
     `ActiveRecord::Base` has it's own `Hash`.